requirements.txt
flask
pyarrow
pandas
azure-storage-blob
fsspec
adlfs
